create initial visualizations
The `visualize` module, which I hadn't planned on prioritizing,
is now a necessary in debugging the model divergences. As such,
this commit creates the intial stages of this module.

It's pretty bare bones, definitely not polished, and not at all
organized the way I want the final product to be. But it will do
for now