add a return distribution option to pulsars
The likelihood function for the pulsar accelerations now has an
option to return the full probability distribution, rather than
just the total likelihood.

This is meant to be used in visualizations initially, but also
I would like to include similar things in all functions, as I
would like to probe and profile a lot more stuff, including the
individual lieklihood functions and their prob. distributions.

This commit also changes the s parameter to s2, because this
value is actually "s^2", I think that should be represented.