change total proper motion likelihood variables
Switches the total proper motion likelihood function to
compare the `PM_tot` values against the pythagorean combination
of the radial and tangential components, rather than the old
`v2j` component, which was clearly wrong.

This may not be completely correct either, see issue #6 for
discussion, but it's better than what we had atleast.

The visualizations are also updated to reflect the changes.