add tracking of individual likelihoods with blobs
After being worried for a while (see #5) about how to track the
individual function likelihoods I've realized that the blobs
(which I kindly ignored before) fit the bill almost exactly.

After some messing around, I think this commit will correctly
track the likelihoods, however this is just a quick update in
order to actually test this hypothesis. If it does work, steps
will need to be taken to ensure it's not slow, or takeing up too
much space