redo handling of nested groups in `data`
The way we decided to handle subgroups before was just not right,
it would have required supplying the subgroup for each call to a
variable, as that is where the name was evaluated.

Now, we don't store any parent groups of groups at all. All data
must exist under one of the lowest-level subgroups in each group.

This will make dealing with annoying situations like 47 Tuc's
proper motion easier, but its a bit ham-handed right now. If this
comes up more often, a more elegant solution will be needed