add support for attributes in `data`
In order to utilize the important metadata stored in hdf file
attributes, in various levels, we needed a wrapper around said
functionality within `data`. To do this in the most logical
sense (in my mind), we made a simple subclass of numpy arrays
which will now represent the variable data.

This also allows us to enforce that the observation data should be
readonly easily.