add dynamic selection of likelihood functions
Closes #4

To facilitate the generalization of fitter to any other cluster,
one of the most important issues was the different likelihood
components which could be evaluated based on available data.

This commit aims to fix this, by introducing a new function in
likelihoods which, based on an Observations, determines which
components are possible to compute, and returns the relevant
info, to be passed to the log_probability function from main.

This commit will tweak the organization of many smaller pieces
as well, is probably filled with bugs, and was written and torn
down multiple times, so it isn't perfect.

It also tries to be as general as possible, but will rely on
*some* data standards, as they are documented