refactor mass function likelihood parameters
This commit simply brings the mass function likelihood up to
date with all the other likelihoods, by restricting the arguments
a bit, and utilizing the relevant fields from the model.

I don't love this design I've adopted, storing extra necessary
fields in the model attributes, but for now it will do.