add support for dynamic tracer star masses
Certain datasets have explicit "tracer" masses, which correspond
to their observation characterstics, which must be given their
own explicit mass bins to inspect in the model, for the
corresponding likelihoods. This commit replaces the old manual
method with a dynamic appending of these bins based on a mass
stored in the metadata of the datasets.

It is optional, and if not provided in the data files, reverts
to using the last main sequence bin (nms-1)