add reading of some imf settings from observations
Some of the parameters passed to `evolve_mf` are cluster
dependant, and should be read from the observation metadata.

More work is required to figure out exatcly which fields are
cluster dependant, and to store them in the relevant data files.
Currently only the metallicity and the 47 Tuc Ndot are stored.