add ability to fix certain parameters
This commit finally adds the functionality of allowing the caller
to supply "fixed" constant values for certain theta parameters.
These parameters will be used in the model creations, but not
varied, used or stored by the MCMC code, thus removing them
from the chains.

This is preliminary, still needs testing and also implementing in
the CLI and visualizers.