add unit tests for data functionalities
This add's one half of the new unit testing (half in scope that
is, not in scale), testing both the functionality of `data` and
all of the available cluster data files for data standard
compliance.

Right now "compliance" just means that they have the required
datasets, however this will probably get more strict in the
future, requiring stuff like specific metadata, correctly
constructed datasets and etc.