change design of fixed parameters functionality
Due to the realization that passing a dictionary of fixed
parameters and their values through argparse would be nigh on
impossible to do (nicely at least), this commit changes said
functionality to instead receive and pass from the user only a
list of parameters to fix. The actual fixed values must be now
provided by the data files, the defaults, or by the actual
`initials` argument.

Passing this argument through argparse is not currently supported
so for the time being it will always fall back on the data file,
but in the future I foresee these values being written into
passed files or something along those lines.

The main function is now becoming overly confusing, and this
commit I think marks the definite need for a complete
documentation overhaul.