vectorize the pulsar probability code
This commit refactors the entirety of `new_Paz.vecPaz` in order
to remove the looping over `A_SPACE` and generalize the function
to other clusters.

The main driver behind this commit was to generalize `A_SPACE`,
however it became clear that in order to do so, a reworking of
the whole apparatus was necessary.

A number of important code changes were made, however the inner
workings and math of the functions should have remained identical.

A good amount of clarification edits are still required, and the
function/likelihood should be also vectorized along the number of
pulsars as well, in the future.