refactor viz dimension parameters as properties
In order to fix in the cleanest fashion the various bugs and
issues plaguing the dimensions of the `RunVisualizer`, due to
weird slice parameters, invalid values and incomplete runs, this
commit completely refactors every part of the dimension logic to
properties, which give us a more strict and validated logic.

This commit also rearranges a couple method orders.