allow a less strict version of `show_obs` in viz
This commit, besides some rearranging and cleanup, allows for
the `show_obs` arguments to all of the model visualizations to
also accept a string 'attempt', which will attempt to plot the
observation data but aso fail gracefully and still plot the model
when that is not possible.

This is designed for when we want to plot the outputs for
multiple models at once, or when we wish to generalize calls to
all the plots for cluster which may not have certain datasets.