redo async error construction
The building of the sometimes asymmetric errors on the dynamics
data is finally moved in `data` where I always said it belonged.

The method is rebuild to be more general, although still will
probably go through some rejigging. The likelihood functions are
also updated.

This also fixs a bug that, while I don't think was affecting
anything currently, could've become a problem at some point, as
the previous async_err function was only valid for LOS
dispersions, and was using `model.v2p` which is not right at all.