redo logging functionality
Logging is changed to better follow what should be expected of
a library: all logging config is left up to the user of said
library.

In our case this user is `GCfitter`, so thats where the config
goes, allowing us to better control it from the args.

This commit also adds a few more logging statements here and
there, cause you can never have enough of those