create a new model class
This commit finally gives a solution to the mess introduced by
the old `create_model`, which I realized made much more sense
as a subclass in the end. `create_model` is removed in favour
of a simple limepy subclass `Model` which creates a limepy base
model from a given theta, just as in the old function, but now
has a much cleaner and standard way to store and carry the other
needed parameters.

Now rather than requiring that our functions be given a limepy
model that had to be created by a specific function (an awful
requirement), we can simply require this class.