change M62 Pdot data key names
This just changes the `dP_meas` key to `Pdot_meas` so that the
new likelihoods will actually work. This change should be made
everywhere (dP -> Pdot), this is just a start.

Also a full cleanup of all data files' pulsar datasets is needed
to work with the new likelihood, this commit does not do that.