implement new log-linear interpolated likelihood
Changes all the new pulsar likelihood function to the even
newer method, based on a log-spaced initial domain, interpolated
onto a dense linear grid for convolution.

Also fixes a critical error in the integration of the intrinsic
distribution. It is now correctly transformed to the same domain
as the distributions it is being convolved with.

This work stems from experimentation in the `Pdot_sandbox` repo,
see there for more information.