Change `Variable` to a subclass of `Quantity`
In order to best handle the switch to astropy units throughout,
the Variable class used by all observational data is changed
to be a subclass of `Quantity` rather than `ndarray`.

There is still some trouble getting the `writeable` flag to work,
so for now it is only "readonly" in name, but it should still be
treated as such.

The class is also taken outside of `Dataset` cause that was just
an ugly way of having it.