implement units in the `Model`
The astropy units are added to various attributes of the in-house
`data.model` subclass. As the original model has many outputs, in not
always the most obvious places, this commit is very much a starting
place. The units are added to the few parameters which I believe
are used in our fitting.

This should be enough to get us up and running but *much* work is
still required on this front. Similarly, the limepy scaling is
not even considered yet, which may be a mistake.