add the binary orbit version of pulsar likelihood
Refactors the old version of the pulsar fitting code we were
originally (without the intrinsic period derivatives) and adds it
as a seperate likelihood which is used if the observations
contains some `Pb` data. Right now this is only 47 Tuc.

The 47 Tuc data file is updated, not only to include the necessary
orbital data, but to slice the isolated and binary data into
seperate groups. (Also the r values are changed from parsecs to
arcmins)