fix bug in pulsar likelihoods with small widths
For some pulsar spin timings (namely in 47 Tuc), the widths, which
come from the Pdot error, are just extremely small. So small that
trying to come up with a linear spacing to cover the gaussian
would just be a losing battle.

To fix this the linear domain has a single point at exactly zero
added. This point will *always* sample the gaussian peak at least
at that one point, and thus for very small widths it will act as
a delta function, rather than a gaussian, and therefore not
have an effect during convolution, but also won't break it.

This *does not* change the fact that the 1e-18 and 5000 factors
should be determined in a better fashion, and that the new_Paz
code should still be tweaked to give a normalized distribution.