rename `likelihoods` module to `probabilities`
I was never happy with the fact that all lieklihood functions
both began with and were under the "likelihood" name. To avoid
repetition, to better describe what belongs in the file, and
to have a more descriptive name be used than "components", the
module is renamed and various references throughout are tweaked.