fix unit bug in pulsar likelihoods
The KDE for the intrinsic Pdot requires a period in seconds, and
will not convert automatically (which is understandable
considering it's scaled and logged before it even gets there).

While we could handle that specific case by converting there, it
turns out it's just easier to convert at the beginning rather than
trying to be naive about units the whole way through.