apply same zero-centred domain trick to Paz domain
The trick we committed before, explicitly using a value of
0 in the middle of the domain in order to turn a very small
gaussian into a delta function, works well, but was only applied
to the dense linear domain. This worked for the most part, but
some widths were actually so small that the orignal logged
domain couldn't even sample it (47 Tuc).

This commit simply applies that trick to the Paz code as well.