create a rough unit-conscious version of `interp1d`
This is functionality that should in principle eventually
exist natively, but for now in order to stop having to re-add
units after each interpolation we've created a ver simple
subclass of Scipy's interp1d which retains units.

The rest of the package is still not consistent in which
interpolation methods it uses, so once that is decided, which
method this is based on might change as well.