add initial confidence interval plotting logic
This commit finally introduces a new version of the
`ModelVisualizer` which handles confidence intervals from a number
of walkers, rather than just plotting the median model.

It is *very* rough and quite ugly, but it works for now.
There's a lot of repetition with the original model viz class
which I think needs to be reduced, but `visualize` itself has
also simply become cluttered.