major restructuring of package into subpackages
As some of the few modules contained directly under `fitter`
became unruly, it was becoming clear that they would need
to be split up into seperate modules. In order to not clutter the
namespace and the fitter package directly, the enire package was
therefore restructured into seperate subpackages for each logical
category of modules.
This should allow us to more easily split apart any of the
current modules we desire, as well as keep the `fitter` namespace
clear as possible