move some run metadata storage to start of runtime
The metadata which doesn't require run results, like fixed params,
is moved to be stored before the run starts. This should ensure
that when a run does not finish, as much metadata as possible is
still stored.
This is important as the current viz logic also sorta breaks
down in these cases.