add functionality for specifying prior bounds
Much in the same way the ability to specify custom values for
initial walker positions, both fixed and otherwise, was added,
this commit adds the ability to specify (as a dict to `fit` and
a JSON file to `GCfitter`) custom bounds on any parameter, to
replace the old static bounds used before.

This involves restructuring `priors` back into a seperate
function (again) and also allows for easy check to be made
against the priors in other locations, like before beginning the
run.

Also added alongside this is the saving of additional metadata
related to these new bounds _and_ the specified intials to the
results file.