ensure pulsar Paz distribution is normalized
The pulsar cluster component was previously not a normalized
distribution, but was sampling the azmax asymptote too closely,
leading to inflated values at the bounds.
This ensures the area under the curve is equal to (almost) 1,
by cutting the domain explicitly.

This also, in turn, eliminates the need for any log spacing in
this function.