add unitless functionality to QuantitySpline
This commit allows x or y with no units to be used in the spline
class.
This was done so I could be consistent with what spline I use
in likelihoods, and this is needed for the mass function
likelihood.

This is a little bit rough, and there are definitely some edge
cases which still need to ironed out.