fix overlapping polygons in a single Field
Fields with multiple polygons, within which some of them overlap
(i.e. the Richer data for 47 Tuc), were failing as the multipolygon
would be overlapping and thus invalid. This fixes that by first
imposing a union, and thus merging those polygons.

This commit also slightly renames some of the 47 Tuc fields,
which were incorrect.