attempt to fix orbital pulsar difficulties
The orbital (binary) pulsars seem to be struggling, and due to
how harsh the pulsar likelihoods are (a single pulsar outside
the bounds disrupts the entire posterior) it's stopping the
MCMC from moving at all.

I believe the issue might be due to the new cutoff introduced by
the normalization changes, as the domain outside the cutoff is
no longer returned as zero. This wouldn't affect something like a
spline, which is why the spin pulsars might be okay, but where we
were directly using this domain for all orbital convolutions,
the distribution was being cut quite short, with no tails at all,
which is giving something like 47 Tuc a lot of difficulty in
fitting all the pulsars.