add hyperparameters to all gaussian likelihoods
In the first attempt to address the overfitting of the mass
functions compared to everything else, here we introduce a new
form of gaussian likelihood, with added scaling marginalized
hyperparameters, as seen in Hobson et al. (2002).

This is simply a replacement of the typical gaussian likelihood
equation, as the hyperparameters can be marginalized analytically.

The pulsars are the only component not yet changed, as that will
require more thinking.