update mass function structure in datafiles
changes how the mass functions are organized from the convoluted
stacked format with field masking to the more obvious subgroup
based on survey format.
Dunno why I did it the other way in the first place.

A couple parts of the `Field` and likelihood code will also
need to be updated.