change mass function to evaluate once each dataset
Rather than evaluating the (hyperparametrized) gaussian on each
radial bin, the likelihood is now gathered and evaluated only
once per likelihood function.
Looking at the math now, I don't believe this will affect any
numbers, but it was first thought of in order to improve the
plotting of the `$\alpha_{eff}$`, and so it's still important.