add `hyperparams` option to all likelihoods
The original gaussian likelihood, without the new hyperparams
is re-added, to `util`, and a kwarg is added to each component
likelihood which allows for the choice between the two.

The option is also added to the pulsar likelihoods, not to be
used, but just so the likelihood calls remain compatible.