More fixes to DM, remove hard coded z domain.
Now we put the signs back into the PdotP domain that we return and we
also use the half-light radius as the bounds for the z domain of the
pulsars. I'm pretty sure pulsars should always be within the rh but
maybe if they've had a recent interaction? Worth double checking with
other clusters.