move extra backend features to a new output class
Most other samplers do not have the concept of file-based
backends, like `emcee` does. We definitely want this feature as
it is integral to our results summary and plotting functionality.

Here we move any extra things we did to this file (on top of what
`emcee` does) to it's own class, as we will most likely reuse
these methods in similar but different ways for other samplers.