add first iteration of nested sampling fitter
Based largely on the original `emcee` fit function, which is here
moved in it's own function, the initial architecture of the
`dynesty` nested sampling fit function is created. This function
is very much not tested, and is still missing all storage of
backend material.

The arguments will also need to be adjusted to match the new
sampler.