add storage of nested sampling results
Expands the `NestedSamplingOutput` class to accomodate basic
file stuff and the naive adding of a results "dictionary",
like those outputted by `dynesty` runs. The idea here is to,
unlike the iterative approach of `emcee`, combine the results
of the sampling runs after each batch, and overwrite the entire
file at that time.