revert the fixed param decorator
Turns out getting the decorator to work with both pickle (for the
pooling) and with how the priors are returned was immensely more
convoluted than it currently is. So for now it seems the best
method to handle all of this will be to copy how it is done in
`posterior` into the priors (with a slight change due to it being
an object).
This (or basically all the top-level probability functions logic)
may change in the future as I feel it's becoming more and more
convoluted with each new feature.