move transformed `Priors` into seperate class
It turns out the differences between the two kinds of `Priors`
we need were greater than expected, due to how all the other
features we've added (fixed params, dependant priors, etc)
interacted with them. It became unfeasible to keep the
prior functionality in a single `Priors` class with a simple
`transform` difference.

Most previous work on making that idea work was discarded for this
change. The actual individual prior functions, however, are fine(ish)
as they are now.