make prior dependancies computation recursive
this wasn't technically needed at the moment, nor really
looked like it was going to be needed anytime soon, but what we
had before had a glaring open vulnerabilitiy for any sort of
chained dependancies. This should hopefully fix that quickly.
It may not be the most efficient way possible.