fix sampler subparser logic
By adding a somewhat "hidden" parent parser for the subparsers
which includes all the shared arguments.
I have no idea if this is the "correct" way of doing this, but it
seems to work for now, atleast better than the last version was.

This also requires that the sampler choice be given immediately
after the cluster, before any other args, which I am on board for.