fix bug with badly formed dependant priors
any invalid priors were returning 0, which is correct for normal
priors, but not for the prior transforms. They should just be
indicating that this is an invalid point in parameter space and
continuing on. To implement that we return nan instead (which
is also the output of calling the ppf outside of [0,1], which is
nice) and then add a simple (and honestly overdue) check of theta
in `posterior` for any non-finite theta values.