add method storing "on-the-fly" sampling results
To ensure our sampler is actually working, it's desirable to not
only wait till a batch is complete and dumped to store results.
Now we will continuously write the results of each sampling
iteration to a "current_batch" group.
It is not as robust as the full results dump, and is mostly just
for tracking the sampling as it proceeds, so it will be reset for
each batch, when the actual results are dumped.