redo internal plotting logic to accomodate masses
Accounting for tracer masses and plotting the correct bins
of the models was impossible the way we were handling the data
and models completely seperately. This commit wraps the two
seperate functions with one single function which can pass the
knowledge of mass bins onto the models before plotting.

How this will affect all our other plotting functions is not yet
clear. I assume it'll break them all, who knows.