restructure specification and corresponding checks
The nested mess we were working with before was a bad idea, and
unnecessarily hard to implement. We switch here to a flatter
format, with three set levels shared by all datasets, and no
recursion. This allows for a few simple functions, which are used
only once per dataset, to test everything we want.