redo iterations/walkers slicing logic in MCMC viz
Will now explicitly allow for callables or arrays as walkers,
and changes all plotting functions to utilize a single function
to slice up whatever arrays according to this logic.

Also removes the ability to pass walkers/iterations to any
functions. The class properties should always be used, to ensure
the proper checks are done.