get rid of all the `strict` logic
This mess was all set up to avoid simply having a couple different
"versions" of a cluster data file, but at this point, with the
support added for "local" clusters (although they weren't exactly
meant to be used for things like "same as that file but without
one data point", they definitely can be) there is no need for any
of this convoluted logic of "strictness" and "confidence", and I'm
happy to get rid of it.

Need to add some documentation, however, on how one could easily
copy and alter slight things from core clusters, to support this.