fix name of `sample` type kwarg
Due to a mixup between the `DynamicNestedSampler` function and the
actualy `DynamicSampler` class, the sampling type has been being
passed in using the `method` kwargs which, in a perfect example
of why `**kwargs` should be used carefully, was being quietly
ignored. The newest version of `dynesty` correctly errors on this
call.