change param plot to resemble traceplot a lot more
The Higson style plot was simply too involved and tricky to set up
right, and I believe the important information is still
represented well with this style plot.
The `dynesty` traceplot was the inspiration behind this style, but
it is tweaked to make possterior estimation clearer compared to the
"chains".