change max logl bound from max(L) to max(L)-Nlive
This is a test change to see if our method for handling the
plateau weight jumps can be fixed by giving it a bit more
leeway in the upper bound; going back to the bound-Nlive
(i.e. the last "worst" point) instead of just the max bound.
This will hopefully allow for our sampler to actually reach the
bound.