change from using max nlive to first nlive for RHS
Some weird stuff was happening with our bounds (I think),
so this change is to set a more (less?) conservative `max_bound`
using only the original nlive.

The thinking behind using any nlive stuff is that if we've gotten
this many points past this bound before, we can do it again, but
I think using the max nlive was incorrect because there's no
guarantee we got *that* many points past the absolute max logl,
only the last logl we used as a bound.