restrict IMF exponent priors to physical values
The uniform prior bounds on all alpha parameters are restricted
to smaller regions, which can be more easily justified through
physical means.

The low-mass components are restricted to an upper bound of the
Salpeter 2.35, which no other IMF formulation is ever higher than,
in the low-mass regime.

The high-mass component has it's upper bound lowered to 4. This was
determined to be slightly above the exponent which, in the lowest
mass possible clusters, will still allow for any remnants. Models
completely devoid of remnants are likely unphysical, and thus an
upper bound of 4 is plenty high. Final fits which result in a3>3.5
should probably be investigated manually anyways, and maybe have
this bound further reduced.