fix PM ratio divide-by-zero bug at edge of models
In some models the velocity fields (v2Tj, v2Rj) are 0 at the very
edge, and thus the computed anisotropy ratio T/R is NaN. This
messed up the interpolation completely, and caused the posterior
to fail.

This fix simpmly sets the NaN ratio to 1. In the future, when the
interpolation methods are improved in the likleihood functions,
this might not be necessary, but this fix will be okay for now
as the models should always be isotropic in the edges anyways.