remove backend "initial batch" logic
The whole idea behind using the generator results was to avoid
messing up the results, until they were combined for each batch.
But in the case of the initial batch, this doesn't make sense,
as there is no pre-existing results to mess up.
So we can simply store the entire on-the-fly results during that
sampling stage.