shuffle some plotting logic to fix CI tracer bins
As identified in #88, the CI plotting of profiles with
tracer masses was broken by the fixes to mass function plotting
last year.

This rearranges some logic to support this once again. Namely,
the `mj` attribute is once again only set to only contain the
mass bins with actually stored profiles. Then the mass function
plotting, which requires the `mj` mass domain, was explicitly
passed the full `mj` array to each slice.