stop acting like we support load for single models
There's no need (afaik) for a whole save/load shebang for a single
model, as the time to compute is miniscule.
I only begrudgingly added it for CI models because they have a
very long init time.

Since they don't exist for single models, they shouldn't be an
option for collections of single models either.