standardized some quantities from the limepy model
Prompted by the confusion surrounding the scaling of `ra`, this
commit finally makes the distinction that you cannot rely on
the theta parameters to be accessible as attributes of `Model`.
This was thought to be the case before, as `s2` was explicitly
reset for that, but it was sloppy and broken for `M` and `ra`,
so I'm just getting rid of it.

This commit now ensures that all quantities are model quantities
and have the correct units, etc.