explicitly make θ params not attributes of Model
As mentioned in previous commit, `theta` values are not
safely accessible as attributes on `Model` anymore, so this
commit simply forces that to be even more true.
No longer are theta values set as attributes in the beginning,
only to be overwritten by limepy. To facilitate this more,
the params are now publically available under `Model.theta`.

Some params may be set as attributes by limepy, but you can no
longer rely on that fact from our end.