remove "statistics" as a required group
The statistics group is also a bit of a leftover, and only gets
used for a single (MCMC only) attribute, so requiring it is a bit
much.
Removing "statistics" is out of the scope of this branch, but for
now treating it the same as we do individual MCMC likelihoods is
fine.