revert previous commit forcing color cycle usage
Turns out there is one case where forcing everything to a specific
colour is useful; in the plotting of the different component
density profiles (cumulative_mass, etc).

This simply reverts the previous commmit and now allows color to
be passed as a kwarg. But color will not sit as an explicit kwarg,
as I really don't want this usage to be commonplace