add check to run analysis for mismatch `GCFIT_DIR`
Basically just to exist as a warning that whatever run you're
trying to analyse might have used a different observation that
what you're about to pair it with.

This is not perfect, and mostly is just to catch the users
attention and let them be careful.
This will not catch everything, especially in the model analysis,
so extra care should be taken there.