Hotfix/gcfit dir logging (#117)
* add storage of `GCFIT_DIR` to run metadata

May be useful in case of the usage of local cluster files. Really
the only way to hopefully be able to identify what might be needed
to recreate this run, save storing a copy of all used observations
of course, which is the entire point of the metadata.

* add check to run analysis for mismatch `GCFIT_DIR`

Basically just to exist as a warning that whatever run you're
trying to analyze might have used a different observation than
what you're about to pair it with.

This is not perfect, and mostly is just to catch the users
attention and let them be careful.
This will not catch everything, especially in the model analysis,
so extra care should be taken there.