rewrite `_get_param` logic to use full chains
Rather than relying on the `parameter_summary` methods of
individual runs and having to use different logic for all three
types of parameters (param, run, model), this commit changes the
logic to be more streamlined by always using the full chains, and
simply grabbing the 1σ percentiles at the end.

This also removes the requirement for `N_simruns` which means the
bootstrapped stat. errors aren't being added, but they're
negligible.