fix some bugs in recently updated data
The recently updated LOS (#140) and MF (#139) had a handful of
bugs in their implementation that are now fixed.

The LOS data used was not actually the most up to date values, and
some "Baumgardt" σ values are changed.

The MF data was incorrectly using a too-low completeness limit
(increased to 70%) and bins with a completeness of exactly 1.000
should have been treated as 0%. In addition, the MF data for
NGC7078 was redone with correct metallicities.