fix outer bin of one ω-Cen mass function dataset
The `mass_function/Cool` dataset contained an outer bin which
had basically zero stars in it. The values and errors were so
small they were actually resulting in divide-by-zero errors.
This fix simply removes that bin. It turns out the geometric
are covered by this field was so small it will have basically
no effect anyways