add support for model `rh` to `CIModelVisualizer`
In the same fashion as 4af905e add `rh` as an attribute to the
CI models. It already existed for the single model, and now that
we have the `force_model` kwarg it doesn't hurt here as well.

The only difference between this and the sampler `rh` will be the
units, which should be automatically applied to this model `rh`.