add a method for recursively mapping an operation
Meant solely for use with the `_get_param_chains` method,
this decorator simply check for any operation symbol and
recursively maps that to two params. This is to allow the easy
plotting of quantities like `ra / rh`.

Also propogates the `force_model` kwarg to the plotting functions
in order to better allow this sort of quantity