change how `_RunAnalysis` inits new observations
Removes all the warning logic surrounding "GCFIT_DIR" in favour of
just reading in the new `restrict_to` metadata from the run file
and using that every time.
Rather than giving way more freedom and shooting a bunch of
warnings, this now just does it the same way and if you want to
do something else you have to do it yourself.