support arbitrary mass bins in `_model_profile`
Accepts a new keyword `mass_bins` which causes the model profiles
for the given mass bins (indices, not masses) to be plotted, in
addition to the bins corresponding to the shown data. This mostly
only applies to single model visualizers, as extra mass bins
aren't stored in CI models.

This may require a version of python >= 3.9, due to dict unions.
If so, we should also change the setup, and change a few other
dict updates to unions throughout