add alternative generators for the models
Based on a few historical versions of DF cluster models (King,
Woolley, etc.) adds easy generator functions which basically
just add a few defaults to the model init (like to `g`).
The current setup is a bit awkward, this may need to be rethought
in the future, but it works.