fix key bug in storing `restrict_to`
`None` cannot be stored in an hdf file attribute, so trying to
store `restrict_to` when it is `None` (i.e. the default) will not
work.
This key is only really read in `analysis.runs` init, where it
defaults to `None` if it is not stored anyways, so only saving it
when it's set to 'core' or 'local' should be fine.

Surprised this was never found, it breaks almost all runs.