redo/modernize all scripts/entrypoint logic
discards the old "scripts" logic used in `setup.py` in favour of
actual entrypoints, which require the console scripts to be an
actual part of the package and callable as a function. This
is the more modern and robust way to handle scripts.

This commit also removes the handful of ancient, broken and unused
scripts in "./bin", which were never in `setup.py` anyways