add a check for "blob" when storing results
The newest versions of `dynesty` come equipped with "blob"
functionality, I believe similar to in `emcee`, which is great.
But, as we have not yet implemented them at all, the default
behaviour seems to be to fill a numpy array with None values. This
is less than ideal, and cannot be stored in an HDF file for output,
as we usually do.

This commit introduces the temporary fix of simply ignoring the
blob items in the results, until we can implement something for them