Hotfix/numdens obs (#171)
* fix a couple number density dataset globs

A couple places in `analysis.model` used a truncated dataset glob
for the number density, which did not have "*" at the end. This
would ignore any nested number density datasets, which is obviously
incorrect. This commit simply fixes those strings.

* fix an explicit number density key in `plot_all`

When trying to get a good lower y limit in `plot_all` we explicitly
called on `obs['number_density'], but this won't always be correct,
and explicit calls like this should not be used.
This commit simply switches to a glob pattern as per usual.