Hotfix/single ellipsoids (#170)
* add support for single ellipsoid sampler bounds

the dynesty bounding distributions allow for the use of single
ellipsoids, and we also support it, but never supported saving its
metadata, and thus couldn't recreate it for `analysis`. This fixes
that.

* add temporary fix for backwards compatibility

Previous runs using a single ellipsoid simply did not save any
relevant metadata, just skipped this step quietly.
For the time being, this simply allows us to still use these old
runs, as long as you don't try to plot the bounds. This should
be removed later.