bring `GCfitter` defaults in line with `main` func
This commit updates a few of the command line argument defaults
to better match those in the core fitter functions, mostly to
be explicitly clear what is being run each time.
This also required setting a couple arguments to values that
I wouldn't ordinarily recommend (dlogz=0.01, for example) because
those arguments are hidden in dictionaries, and can't nicely have
the defaults I would want set in the functions, so instead revert
to the `dynesty` defaults, which are geared for a bit nicer
distributions, I think.
But given that you should probably supply those args when running
anyways, it's not a big deal.